requirements.txt
# v0agent requirements
# Cache bust: 2026-01-25-v2 force rebuild
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
python-dotenv==1.2.1
openai==2.15.0
python-multipart==0.0.21
# Configuration management
PyYAML==6.0.1
pydantic>=2.0.0

# Embeddings and vector search
chromadb==0.4.24
# Pin numpy <2.0 for chromadb compatibility (np.float_ removed in NumPy 2.0)
numpy<2.0

# Encryption
cryptography==42.0.0

# LangSmith Tracing (for agent observability)
langsmith>=0.1.0
langchain-core>=0.3.0

# Multi-device sync
zeroconf==0.130.0

# Supabase client (optional, for cloud sync)
# Pin specific versions to avoid proxy argument incompatibility
# supabase 2.3.0 requires: postgrest>=0.10.8,<0.14.0
supabase==2.3.0
gotrue==2.4.1
postgrest==0.13.0

# SQLite extensions
# pysqlite3-binary - not needed, using standard sqlite3
# Redis for caching and task queue
redis>=5.0.0
rq>=1.15.0

# HTTP client
requests>=2.31.0
# httpx for async HTTP requests - pin version compatible with supabase 2.3.0
httpx==0.24.1

# Background job scheduler
APScheduler>=3.10.0
anthropic>=0.45.0
PyPDF2>=3.0.0
